fix(app-website): make main rendering components composable
