+++ packages/app-website/src/index.tsx
@@ -1,6 +1,18 @@
+export * from "@webiny/app";
+export type {
+    HigherOrderComponent,
+    ProviderProps,
+    ComposeProps,
+    ComposableFC,
+    AppProps
+} from "@webiny/app";
+
 export * from "@webiny/app-theme";
 export * from "./Website";
 export * from "./Page";
+export * from "./Page/PageRenderer";
+export * from "./Page/WebsiteScripts";
+export * from "./Page/ErrorPage";
 export * from "./Menu";
 export * from "./utils/isPrerendering";
 export * from "./utils/getPrerenderId";
